requirements.txt
asgiref==3.8.1
beem-africa==0.1.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.2
defusedxml==0.7.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
icalendar==6.1.3
idna==3.10
oauthlib==3.2.2
pillow==10.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
